Reworked to have a top level ORM handle the building of queries in a standardized way, then the driver can build it using the data stored by the top level ORM once the query is being executed.
This makes it so that the drivers can focus strictly on what they have to and I'll be able to check back with the query being build no matter what driver is being used.

Currently only works with read queries, model is not working and removed joins for the time being.